Cambié el cierre de sección al servicio de paseo
