+++ fritri-app/src/interfaces/paseo.tsx
@@ -106,3 +106,8 @@ export interface IPaseoUpdate extends IPaseo {
     idPaseo?: string;
     modificacionesRealizadas: string[];
 }
+
+export enum TipoSeccion {
+    RESTAURANTE = 'RESTAURANTE',
+    ATRACCION_TURISTICA = 'ATRACCION_TURISTICA' 
+  }
