+++ paseos-servicio/src/paseos/dto/cerrar-seccion.ts
@@ -1,0 +1,17 @@
+import { IsNotEmpty } from 'class-validator';
+import { TipoSeccion } from '../schemas/paseos.schema';
+
+export class CerrarSeccionDto {
+  
+  @IsNotEmpty()
+  readonly idPaseo: string;
+
+  @IsNotEmpty()
+  readonly tipoSeccion: TipoSeccion;
+
+  @IsNotEmpty()
+  readonly cerrarVotaciones: boolean;
+
+  @IsNotEmpty()
+  readonly fechaModificacion: Date;
+}
