+++ paseos-servicio/src/paseos/paseos.service.ts
@@ -2,13 +2,14 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { BadRequestException } from '@nestjs/common';
-import { Paseo, PaseoDocument } from './schemas/paseos.schema';
+import { Paseo, PaseoDocument, TipoSeccion } from './schemas/paseos.schema';
 import { CrearPaseoDto } from './dto/crear-paseo.dto';
 import { ActualizarPaseoDto } from './dto/actualizar-paseo.dto';
 import { GenerarPinProteger } from '../helpers/generador-codigo';
 import { NotificacionesService } from 'src/notificaciones/notificaciones.service';
 import { NotificacionPaseoActualizado } from 'src/notificaciones/dto/notificacion-paseo-actualizado.dto';
 import { Notificacion } from 'src/notificaciones/dto/notificacion.dto';
+import { CerrarSeccionDto } from './dto/cerrar-seccion';
 
 export enum EstadoPaseo {
   Pendiente = 1,
@@ -206,6 +207,51 @@ export class PaseosService {
       );
     }
     return resultado;
+  }
+
+  async cerrarSeccion(cerrarSeccionDto: CerrarSeccionDto) {
+    let resultadoPaseo;
+    try {
+      const { idPaseo, tipoSeccion, cerrarVotaciones, fechaModificacion } = cerrarSeccionDto;
+      resultadoPaseo = await this.paseoModel.findOne({ _id: idPaseo });
+      if(!resultadoPaseo) {
+        throw new Error(`No existe el paseo con el id::${idPaseo}`);
+      }
+
+      let filter;
+      let update
+      let seccion;
+      if (tipoSeccion===TipoSeccion.RESTAURANTE) {
+        seccion = "Restaurantes";
+        filter = { _id: idPaseo };
+        update = { 'seccionRestaurantes.esFinalizadasVotaciones':cerrarVotaciones, 
+                    'seccionRestaurantes.fechaFinalizacionVotaciones': fechaModificacion}; 
+      } else {
+        seccion = "Atracciones Turísticas";
+        filter = { _id: idPaseo };
+        update = { 'seccionAtraccionesTuristicas.esFinalizadasVotaciones':cerrarVotaciones, 
+                    'seccionAtraccionesTuristicas.fechaFinalizacionVotaciones': fechaModificacion};
+      }
+      
+      resultadoPaseo = await this.paseoModel.findOneAndUpdate(filter,update, {
+        returnOriginal: false
+      });  
+
+      const notificacionPaseoActualizado: NotificacionPaseoActualizado = {
+        idPaseo: idPaseo,
+        nombrePaseo: resultadoPaseo.nombre,
+        integrantes: resultadoPaseo.integrantes,
+        modificacionesRealizadas: [ `Se cerraron las votaciones en la sección de ${seccion}.` ]
+      };
+
+      console.log("Notificacion Paseo: " + JSON.stringify(notificacionPaseoActualizado));
+
+      await this.notificarPaseoActualizado(notificacionPaseoActualizado);
+
+    } catch(error) {
+      throw new BadRequestException(`Error al tratar de cerrar sección de votos::${error.message}`);
+    }
+    return resultadoPaseo;
   }
 
   async eliminar(idPaseo: string) {
