+++ paseos-servicio/src/votaciones/dto/cerrar-seccion.ts
@@ -1,17 +1,17 @@
 import { IsNotEmpty } from 'class-validator';
 import { TipoSeccion } from '../interface/votar-seccion';
 
-export class CerrarSeccionDto {
+// export class CerrarSeccionDto {
   
-  @IsNotEmpty()
-  readonly idPaseo: string;
+//   @IsNotEmpty()
+//   readonly idPaseo: string;
 
-  @IsNotEmpty()
-  readonly tipoSeccion: TipoSeccion;
+//   @IsNotEmpty()
+//   readonly tipoSeccion: TipoSeccion;
 
-  @IsNotEmpty()
-  readonly cerrarVotaciones: boolean;
+//   @IsNotEmpty()
+//   readonly cerrarVotaciones: boolean;
 
-  @IsNotEmpty()
-  readonly fechaModificacion: Date;
-}
+//   @IsNotEmpty()
+//   readonly fechaModificacion: Date;
+// }
