+++ paseos-servicio/src/votaciones/votaciones.controller.ts
@@ -1,5 +1,5 @@
 import { Body, Controller, Delete, Get, HttpCode, Param, Patch, Post, Put, Query, UsePipes, ValidationPipe } from '@nestjs/common';
-import { CerrarSeccionDto } from './dto/cerrar-seccion';
+//import { CerrarSeccionDto } from './dto/cerrar-seccion';
 import { VotarSeccionDto } from './dto/votar-seccion';
 import { VotacionesService } from './votaciones.service';
 
@@ -14,12 +14,12 @@ export class VotacionesController {
     return await this.votacionesService.votarSeccion(votarSeccionDto);
   }
 
-  @Patch('cerrar-seccion')
-  @UsePipes(new ValidationPipe())
-  @HttpCode(200)
-  async cerrarSeccion(@Body() cerrarSeccionDto: CerrarSeccionDto) {
-    return await this.votacionesService.cerrarSeccion(cerrarSeccionDto);
-  }
+  // @Patch('cerrar-seccion')
+  // @UsePipes(new ValidationPipe())
+  // @HttpCode(200)
+  // async cerrarSeccion(@Body() cerrarSeccionDto: CerrarSeccionDto) {
+  //   return await this.votacionesService.cerrarSeccion(cerrarSeccionDto);
+  // }
 
 
   @Post('quitar-voto-seccion')
